requirements.txt
# Core HTTP and Async Libraries
aiohttp==3.9.1
requests==2.31.0

# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning (for ITF match predictor)
scikit-learn==1.3.2
joblib==1.3.2

# Configuration and Environment
PyYAML==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Browser Automation
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1
fake-useragent==1.4.0

# Utilities
click==8.1.7
tqdm==4.66.1
colorama==0.4.6
schedule==1.2.0

# OpenAI Integration (for AI analyzer)
openai==1.3.7
tiktoken==0.5.2

# Notion Integration
notion-client==2.2.1

# Telegram Bot (for notifications)
python-telegram-bot==20.7
pytz==2023.3

# Development and Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
